move models.items to core.models
