+++ lenny/core/__init__.py
@@ -1,15 +1,37 @@
 #!/usr/bin/env python
 
 """
-    Core module for Lenny,
-    including the main application setup and configuration.
+    Core module for Lenny, s3 & db
     
     :copyright: (c) 2015 by AUTHORS
     :license: see LICENSE for more details
 """
 
 import boto3
-from lenny.configs import S3_CONFIG
+from sqlalchemy import create_engine
+from sqlalchemy.orm import scoped_session, sessionmaker, declarative_base 
+from lenny.configs import DB_URI, DEBUG, S3_CONFIG
+
+Base = declarative_base()
+
+# Import all models here to ensure they are registered with Base
+from lenny.core.models import Item
+
+# Configure Database Connection
+engine = create_engine(DB_URI, echo=DEBUG, client_encoding='utf8')
+db = scoped_session(sessionmaker(bind=engine, autocommit=False, autoflush=False))
+
+def init_db(engine_to_init=engine):
+    """Initializes the database and creates tables."""
+    Base.metadata.create_all(bind=engine_to_init)
+
+def _auto_init_db():
+    try:
+        init_db(engine)
+    except Exception as e:
+        print(f"[WARNING] Database initialization failed: {e}")
+
+_auto_init_db()
 
 class LennyS3:
 
@@ -57,4 +79,5 @@ class LennyS3:
 
 s3 = LennyS3()
                 
-__all__ = ['s3']
+
+__all__ = ["s3", "Base", "db", "engine", "items", "init_db"]
